Refactor package structure for constants to improve organization
