+++ app/src/main/java/com/davidread/habittracker/common/di/CommonModule.kt
@@ -1,7 +1,7 @@
 package com.davidread.habittracker.common.di
 
 import android.app.Application
-import com.davidread.habittracker.common.constant.BaseUrl
+import com.davidread.habittracker.common.config.BaseUrl
 import com.davidread.habittracker.common.network.AuthenticationInterceptor
 import com.davidread.habittracker.common.repository.AuthenticationTokenRepository
 import com.davidread.habittracker.common.repository.PREFS_NAME
